Redid a bit of the logic on LoadPifRom, it will now fail sanely instead of expecting results to be perfect Also went over the file as a whole and added more whitespace for an easier read

+++ Pif.h
@@ -26,6 +26,6 @@
 extern BYTE PifRom[0x7C0], *PIF_Ram;
 
 int  GetCicChipID ( BYTE *RomData );
-int  LoadPifRom   ( int country );
+int  LoadPifRom   ( BYTE country );
 void PifRamWrite  ( void );
 void PifRamRead   ( void );
